Fixes warning when pressing right arrow key while focusing the minute's input.
Adds a check on the time picker so it doesn't try to focus the AM/PM
input whenever the format is set to 24hr.
Changes check on the onChange() event so it does not change the time if
there is no AM/PM selected on 12hr format and it does change the time
on 24hr format without AM/PM selected.