feat: add `hour24` to the TimePicker component (#1230)
* Adds hour24 prop to time picker component.
Time picker component can receive a prop called hour24 which sets the
component to a 24hr format...

* Fixes time picker for 24hr format.
Enables time to be picked in a 24hr format. Also adds an example for
the hour24 prop.

* Fixes warning when pressing right arrow key while focusing the minute's
input.

Adds a check on the time picker so it doesn't try to focus the AM/PM
input whenever the format is set to 24hr.
Changes check on the onChange() event so it does not change the time if
there is no AM/PM selected on 12hr format and it does change the time
on 24hr format without AM/PM selected.